Renames FeathersApp to FeathersAppProvider
