+++ src/components/FeathersAppProvider.js
